+++ lib/bootstrap_form.rb
@@ -1,4 +1,5 @@
 require 'bootstrap_form/form_builder'
+require 'bootstrap_form/form_builder_form_with'
 require 'bootstrap_form/helper'
 
 module BootstrapForm