+++ test/dummy/db/schema.rb
@@ -18,8 +18,8 @@ ActiveRecord::Schema.define(version: 20140922133133) do
     t.string "city"
     t.string "state"
     t.string "zip_code"
-    t.datetime "created_at"
-    t.datetime "updated_at"
+    t.datetime "created_at", null: false
+    t.datetime "updated_at", null: false
   end
 
   create_table "users", force: :cascade do |t|
@@ -28,8 +28,8 @@ ActiveRecord::Schema.define(version: 20140922133133) do
     t.text "comments"
     t.string "status"
     t.string "misc"
-    t.datetime "created_at"
-    t.datetime "updated_at"
+    t.datetime "created_at", null: false
+    t.datetime "updated_at", null: false
     t.text "preferences"
     t.boolean "terms", default: false
     t.string "type"